refactoring as requested through the review process.
