+++ src/main/kotlin/org/openbase/bco/device/hass/manager/dto/HassFloorDto.kt
@@ -2,7 +2,7 @@ package org.openbase.bco.device.hass.manager.dto
 
 import com.google.gson.annotations.SerializedName
 import org.openbase.bco.device.hass.type.InputDtoProvider
-import org.openbase.bco.device.hass.util.Mergeable
+import org.openbase.bco.device.hass.type.Mergeable
 
 /**
  * Original Type:
