+++ src/main/kotlin/org/openbase/bco/device/hass/type/Mergeable.kt
@@ -1,4 +1,4 @@
-package org.openbase.bco.device.hass.util
+package org.openbase.bco.device.hass.type
 
 import org.openbase.bco.device.hass.manager.dto.HassDto
 import org.openbase.bco.device.hass.manager.dto.HassInputDto
